turn RLPFormat into multiple RLPFormats for each hard fork
