+++ ethereum/core/src/main/java/org/hyperledger/besu/ethereum/encoding/FrontierRLPFormat.java
@@ -1,0 +1,139 @@
+/*
+ *
+ *  * Copyright ConsenSys AG.
+ *  *
+ *  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ *  * the License. You may obtain a copy of the License at
+ *  *
+ *  * http://www.apache.org/licenses/LICENSE-2.0
+ *  *
+ *  * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ *  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ *  * specific language governing permissions and limitations under the License.
+ *  *
+ *  * SPDX-License-Identifier: Apache-2.0
+ *
+ */
+
+package org.hyperledger.besu.ethereum.encoding;
+
+import org.apache.tuweni.bytes.Bytes;
+import org.hyperledger.besu.config.GoQuorumOptions;
+import org.hyperledger.besu.crypto.SECP256K1;
+import org.hyperledger.besu.ethereum.core.Address;
+import org.hyperledger.besu.ethereum.core.Transaction;
+import org.hyperledger.besu.ethereum.core.Wei;
+import org.hyperledger.besu.ethereum.rlp.RLPInput;
+import org.hyperledger.besu.ethereum.rlp.RLPOutput;
+
+import java.math.BigInteger;
+import java.util.Optional;
+
+import static org.hyperledger.besu.ethereum.core.Transaction.GO_QUORUM_PRIVATE_TRANSACTION_V_VALUE_MAX;
+import static org.hyperledger.besu.ethereum.core.Transaction.GO_QUORUM_PRIVATE_TRANSACTION_V_VALUE_MIN;
+import static org.hyperledger.besu.ethereum.core.Transaction.REPLAY_PROTECTED_V_BASE;
+import static org.hyperledger.besu.ethereum.core.Transaction.REPLAY_PROTECTED_V_MIN;
+import static org.hyperledger.besu.ethereum.core.Transaction.REPLAY_UNPROTECTED_V_BASE;
+import static org.hyperledger.besu.ethereum.core.Transaction.REPLAY_UNPROTECTED_V_BASE_PLUS_1;
+import static org.hyperledger.besu.ethereum.core.Transaction.TWO;
+
+public class FrontierRLPFormat implements RLPFormat {
+
+  @Override
+  public void encode(final Transaction transaction, final RLPOutput rlpOutput) {
+    rlpOutput.startList();
+    rlpOutput.writeLongScalar(transaction.getNonce());
+    rlpOutput.writeUInt256Scalar(transaction.getGasPrice());
+    rlpOutput.writeLongScalar(transaction.getGasLimit());
+    rlpOutput.writeBytes(transaction.getTo().map(Bytes::copy).orElse(Bytes.EMPTY));
+    rlpOutput.writeUInt256Scalar(transaction.getValue());
+    rlpOutput.writeBytes(transaction.getPayload());
+    writeSignature(transaction, rlpOutput);
+    rlpOutput.endList();
+  }
+
+  static void writeSignature(final Transaction transaction, final RLPOutput out) {
+    out.writeBigIntegerScalar(transaction.getV());
+    out.writeBigIntegerScalar(transaction.getSignature().getR());
+    out.writeBigIntegerScalar(transaction.getSignature().getS());
+  }
+
+  @Override
+  public Transaction decode(final RLPInput rlpInput) {
+    if (GoQuorumOptions.goquorumCompatibilityMode) {
+      return decodeGoQuorum(rlpInput);
+    }
+    rlpInput.enterList();
+    final Transaction.Builder builder =
+        Transaction.builder()
+            .nonce(rlpInput.readLongScalar())
+            .gasPrice(Wei.of(rlpInput.readUInt256Scalar()))
+            .gasLimit(rlpInput.readLongScalar())
+            .to(rlpInput.readBytes(v -> v.size() == 0 ? null : Address.wrap(v)))
+            .value(Wei.of(rlpInput.readUInt256Scalar()))
+            .payload(rlpInput.readBytes());
+
+    final BigInteger v = rlpInput.readBigIntegerScalar();
+    final byte recId;
+    Optional<BigInteger> chainId = Optional.empty();
+    if (v.equals(REPLAY_UNPROTECTED_V_BASE) || v.equals(REPLAY_UNPROTECTED_V_BASE_PLUS_1)) {
+      recId = v.subtract(REPLAY_UNPROTECTED_V_BASE).byteValueExact();
+    } else if (v.compareTo(REPLAY_PROTECTED_V_MIN) > 0) {
+      chainId = Optional.of(v.subtract(REPLAY_PROTECTED_V_BASE).divide(TWO));
+      recId = v.subtract(TWO.multiply(chainId.get()).add(REPLAY_PROTECTED_V_BASE)).byteValueExact();
+    } else {
+      throw new RuntimeException(
+          String.format("An unsupported encoded `v` value of %s was found", v));
+    }
+    final BigInteger r = rlpInput.readUInt256Scalar().toBytes().toUnsignedBigInteger();
+    final BigInteger s = rlpInput.readUInt256Scalar().toBytes().toUnsignedBigInteger();
+    final SECP256K1.Signature signature = SECP256K1.Signature.create(r, s, recId);
+
+    rlpInput.leaveList();
+
+    chainId.ifPresent(builder::chainId);
+    return builder.signature(signature).build();
+  }
+
+  static Transaction decodeGoQuorum(final RLPInput input) {
+    input.enterList();
+
+    final Transaction.Builder builder =
+        Transaction.builder()
+            .nonce(input.readLongScalar())
+            .gasPrice(Wei.of(input.readUInt256Scalar()))
+            .gasLimit(input.readLongScalar())
+            .to(input.readBytes(v -> v.size() == 0 ? null : Address.wrap(v)))
+            .value(Wei.of(input.readUInt256Scalar()))
+            .payload(input.readBytes());
+
+    final BigInteger v = input.readBigIntegerScalar();
+    final byte recId;
+    Optional<BigInteger> chainId = Optional.empty();
+    if (isGoQuorumPrivateTransaction(v)) {
+      // GoQuorum private TX. No chain ID. Preserve the v value as provided.
+      builder.v(v);
+      recId = v.subtract(GO_QUORUM_PRIVATE_TRANSACTION_V_VALUE_MIN).byteValueExact();
+    } else if (v.equals(REPLAY_UNPROTECTED_V_BASE) || v.equals(REPLAY_UNPROTECTED_V_BASE_PLUS_1)) {
+      recId = v.subtract(REPLAY_UNPROTECTED_V_BASE).byteValueExact();
+    } else if (v.compareTo(REPLAY_PROTECTED_V_MIN) > 0) {
+      chainId = Optional.of(v.subtract(REPLAY_PROTECTED_V_BASE).divide(TWO));
+      recId = v.subtract(TWO.multiply(chainId.get()).add(REPLAY_PROTECTED_V_BASE)).byteValueExact();
+    } else {
+      throw new RuntimeException(
+          String.format("An unsupported encoded `v` value of %s was found", v));
+    }
+    final BigInteger r = input.readUInt256Scalar().toBytes().toUnsignedBigInteger();
+    final BigInteger s = input.readUInt256Scalar().toBytes().toUnsignedBigInteger();
+    final SECP256K1.Signature signature = SECP256K1.Signature.create(r, s, recId);
+
+    input.leaveList();
+    chainId.ifPresent(builder::chainId);
+    return builder.signature(signature).build();
+  }
+
+  private static boolean isGoQuorumPrivateTransaction(final BigInteger v) {
+    return v.equals(GO_QUORUM_PRIVATE_TRANSACTION_V_VALUE_MAX)
+        || v.equals(GO_QUORUM_PRIVATE_TRANSACTION_V_VALUE_MIN);
+  }
+}
