+++ ethereum/core/src/main/java/org/hyperledger/besu/ethereum/encoding/RLPFormat.java
@@ -1,251 +1,38 @@
 /*
- * Copyright ConsenSys AG.
  *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
+ *  * Copyright ConsenSys AG.
+ *  *
+ *  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ *  * the License. You may obtain a copy of the License at
+ *  *
+ *  * http://www.apache.org/licenses/LICENSE-2.0
+ *  *
+ *  * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ *  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ *  * specific language governing permissions and limitations under the License.
+ *  *
+ *  * SPDX-License-Identifier: Apache-2.0
  *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
- * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations under the License.
- *
- * SPDX-License-Identifier: Apache-2.0
  */
+
 package org.hyperledger.besu.ethereum.encoding;
 
-import org.hyperledger.besu.config.GoQuorumOptions;
-import org.hyperledger.besu.config.experimental.ExperimentalEIPs;
-import org.hyperledger.besu.crypto.SECP256K1;
-import org.hyperledger.besu.ethereum.core.Address;
 import org.hyperledger.besu.ethereum.core.Transaction;
-import org.hyperledger.besu.ethereum.core.Wei;
-import org.hyperledger.besu.ethereum.rlp.RLP;
 import org.hyperledger.besu.ethereum.rlp.RLPInput;
 import org.hyperledger.besu.ethereum.rlp.RLPOutput;
-import org.hyperledger.besu.plugin.data.Quantity;
-import org.hyperledger.besu.plugin.data.TransactionType;
 
-import java.math.BigInteger;
-import java.util.Optional;
-
-import com.google.common.collect.ImmutableMap;
-import org.apache.tuweni.bytes.Bytes;
-
-import static org.hyperledger.besu.ethereum.core.Transaction.GO_QUORUM_PRIVATE_TRANSACTION_V_VALUE_MAX;
-import static org.hyperledger.besu.ethereum.core.Transaction.GO_QUORUM_PRIVATE_TRANSACTION_V_VALUE_MIN;
-import static org.hyperledger.besu.ethereum.core.Transaction.REPLAY_PROTECTED_V_BASE;
-import static org.hyperledger.besu.ethereum.core.Transaction.REPLAY_PROTECTED_V_MIN;
-import static org.hyperledger.besu.ethereum.core.Transaction.REPLAY_UNPROTECTED_V_BASE;
-import static org.hyperledger.besu.ethereum.core.Transaction.REPLAY_UNPROTECTED_V_BASE_PLUS_1;
-import static org.hyperledger.besu.ethereum.core.Transaction.TWO;
-
-public class RLPFormat {
-
+public interface RLPFormat {
   @FunctionalInterface
-  interface Encoder {
-    void encode(Transaction transaction, RLPOutput output);
-  }
-
-  private static final ImmutableMap<TransactionType, Encoder> TYPED_TRANSACTION_ENCODERS =
-      ImmutableMap.of(TransactionType.EIP1559, RLPFormat::encodeEIP1559);
-
-  public static void encode(final Transaction transaction, final RLPOutput rlpOutput) {
-    if (transaction.getType().equals(TransactionType.FRONTIER)) {
-      encodeFrontier(transaction, rlpOutput);
-    } else {
-      final TransactionType type = transaction.getType();
-      final Encoder encoder =
-          Optional.ofNullable(TYPED_TRANSACTION_ENCODERS.get(type))
-              .orElseThrow(
-                  () ->
-                      new IllegalStateException(
-                          String.format(
-                              "Developer Error. A supported transaction type %s has no associated"
-                                  + " encoding logic",
-                              type)));
-      rlpOutput.writeRaw(
-          Bytes.concatenate(
-              Bytes.of((byte) type.getSerializedType()),
-              RLP.encode(output -> encoder.encode(transaction, output))));
-    }
-  }
-
-  static void encodeFrontier(final Transaction transaction, final RLPOutput out) {
-    out.startList();
-    out.writeLongScalar(transaction.getNonce());
-    out.writeUInt256Scalar(transaction.getGasPrice());
-    out.writeLongScalar(transaction.getGasLimit());
-    out.writeBytes(transaction.getTo().map(Bytes::copy).orElse(Bytes.EMPTY));
-    out.writeUInt256Scalar(transaction.getValue());
-    out.writeBytes(transaction.getPayload());
-    writeSignature(transaction, out);
-    out.endList();
-  }
-
-  static void encodeEIP1559(final Transaction transaction, final RLPOutput out) {
-    ExperimentalEIPs.eip1559MustBeEnabled();
-
-    out.startList();
-    out.writeLongScalar(transaction.getNonce());
-    out.writeNull();
-    out.writeLongScalar(transaction.getGasLimit());
-    out.writeBytes(transaction.getTo().map(Bytes::copy).orElse(Bytes.EMPTY));
-    out.writeUInt256Scalar(transaction.getValue());
-    out.writeBytes(transaction.getPayload());
-    out.writeUInt256Scalar(
-        transaction.getGasPremium().map(Quantity::getValue).map(Wei::ofNumber).orElseThrow());
-    out.writeUInt256Scalar(
-        transaction.getFeeCap().map(Quantity::getValue).map(Wei::ofNumber).orElseThrow());
-    writeSignature(transaction, out);
-    out.endList();
-  }
-
-  private static void writeSignature(final Transaction transaction, final RLPOutput out) {
-    out.writeBigIntegerScalar(transaction.getV());
-    out.writeBigIntegerScalar(transaction.getSignature().getR());
-    out.writeBigIntegerScalar(transaction.getSignature().getS());
+  interface Encoder<T> {
+    void encode(T object, RLPOutput output);
   }
 
   @FunctionalInterface
-  interface Decoder {
-    Transaction decode(RLPInput input);
+  interface Decoder<T> {
+    T decode(RLPInput input);
   }
 
-  private static final ImmutableMap<TransactionType, Decoder> TYPED_TRANSACTION_DECODERS =
-      ImmutableMap.of(TransactionType.EIP1559, RLPFormat::decodeEIP1559);
+  void encode(Transaction transaction, RLPOutput rlpOutput);
 
-  public static Transaction decode(final RLPInput rlpInput) {
-    if (GoQuorumOptions.goquorumCompatibilityMode) {
-      return decodeGoQuorum(rlpInput);
-    }
-    final Bytes typedTransactionBytes = rlpInput.raw();
-    final int firstByte = typedTransactionBytes.get(0) & 0xff;
-    final TransactionType transactionType = TransactionType.of(firstByte);
-    if (transactionType.equals(TransactionType.FRONTIER)) {
-      return decodeFrontier(rlpInput);
-    } else {
-      rlpInput.skipNext(); // throw away the type byte
-      final Decoder decoder =
-          Optional.ofNullable(TYPED_TRANSACTION_DECODERS.get(transactionType))
-              .orElseThrow(
-                  () ->
-                      new IllegalStateException(
-                          String.format(
-                              "Developer Error. A supported transaction type %s has no associated"
-                                  + " decoding logic",
-                              transactionType)));
-      return decoder.decode(rlpInput);
-    }
-  }
-
-  static Transaction decodeFrontier(final RLPInput input) {
-    input.enterList();
-    final Transaction.Builder builder =
-        Transaction.builder()
-            .nonce(input.readLongScalar())
-            .gasPrice(Wei.of(input.readUInt256Scalar()))
-            .gasLimit(input.readLongScalar())
-            .to(input.readBytes(v -> v.size() == 0 ? null : Address.wrap(v)))
-            .value(Wei.of(input.readUInt256Scalar()))
-            .payload(input.readBytes());
-
-    final BigInteger v = input.readBigIntegerScalar();
-    final byte recId;
-    Optional<BigInteger> chainId = Optional.empty();
-    if (v.equals(REPLAY_UNPROTECTED_V_BASE) || v.equals(REPLAY_UNPROTECTED_V_BASE_PLUS_1)) {
-      recId = v.subtract(REPLAY_UNPROTECTED_V_BASE).byteValueExact();
-    } else if (v.compareTo(REPLAY_PROTECTED_V_MIN) > 0) {
-      chainId = Optional.of(v.subtract(REPLAY_PROTECTED_V_BASE).divide(TWO));
-      recId = v.subtract(TWO.multiply(chainId.get()).add(REPLAY_PROTECTED_V_BASE)).byteValueExact();
-    } else {
-      throw new RuntimeException(
-          String.format("An unsupported encoded `v` value of %s was found", v));
-    }
-    final BigInteger r = input.readUInt256Scalar().toBytes().toUnsignedBigInteger();
-    final BigInteger s = input.readUInt256Scalar().toBytes().toUnsignedBigInteger();
-    final SECP256K1.Signature signature = SECP256K1.Signature.create(r, s, recId);
-
-    input.leaveList();
-
-    chainId.ifPresent(builder::chainId);
-    return builder.signature(signature).build();
-  }
-
-  static Transaction decodeEIP1559(final RLPInput input) {
-    ExperimentalEIPs.eip1559MustBeEnabled();
-    input.enterList();
-
-    final Transaction.Builder builder =
-        Transaction.builder()
-            .nonce(input.readLongScalar())
-            .gasPrice(Wei.of(input.readUInt256Scalar()))
-            .gasLimit(input.readLongScalar())
-            .to(input.readBytes(v -> v.size() == 0 ? null : Address.wrap(v)))
-            .value(Wei.of(input.readUInt256Scalar()))
-            .payload(input.readBytes())
-            .gasPremium(Wei.of(input.readBytes().toBigInteger()))
-            .feeCap(Wei.of(input.readBytes().toBigInteger()));
-
-    final BigInteger v = input.readBytes().toBigInteger();
-    final BigInteger r = input.readUInt256Scalar().toBytes().toUnsignedBigInteger();
-    final BigInteger s = input.readUInt256Scalar().toBytes().toUnsignedBigInteger();
-    final byte recId;
-    Optional<BigInteger> chainId = Optional.empty();
-    if (v.equals(REPLAY_UNPROTECTED_V_BASE) || v.equals(REPLAY_UNPROTECTED_V_BASE_PLUS_1)) {
-      recId = v.subtract(REPLAY_UNPROTECTED_V_BASE).byteValueExact();
-    } else if (v.compareTo(REPLAY_PROTECTED_V_MIN) > 0) {
-      chainId = Optional.of(v.subtract(REPLAY_PROTECTED_V_BASE).divide(TWO));
-      recId = v.subtract(TWO.multiply(chainId.get()).add(REPLAY_PROTECTED_V_BASE)).byteValueExact();
-    } else {
-      throw new RuntimeException(
-          String.format("An unsupported encoded `v` value of %s was found", v));
-    }
-    final SECP256K1.Signature signature = SECP256K1.Signature.create(r, s, recId);
-    input.leaveList();
-    chainId.ifPresent(builder::chainId);
-    return builder.signature(signature).build();
-  }
-
-  static Transaction decodeGoQuorum(final RLPInput input) {
-    input.enterList();
-
-    final Transaction.Builder builder =
-        Transaction.builder()
-            .nonce(input.readLongScalar())
-            .gasPrice(Wei.of(input.readUInt256Scalar()))
-            .gasLimit(input.readLongScalar())
-            .to(input.readBytes(v -> v.size() == 0 ? null : Address.wrap(v)))
-            .value(Wei.of(input.readUInt256Scalar()))
-            .payload(input.readBytes());
-
-    final BigInteger v = input.readBigIntegerScalar();
-    final byte recId;
-    Optional<BigInteger> chainId = Optional.empty();
-    if (isGoQuorumPrivateTransaction(v)) {
-      // GoQuorum private TX. No chain ID. Preserve the v value as provided.
-      builder.v(v);
-      recId = v.subtract(GO_QUORUM_PRIVATE_TRANSACTION_V_VALUE_MIN).byteValueExact();
-    } else if (v.equals(REPLAY_UNPROTECTED_V_BASE) || v.equals(REPLAY_UNPROTECTED_V_BASE_PLUS_1)) {
-      recId = v.subtract(REPLAY_UNPROTECTED_V_BASE).byteValueExact();
-    } else if (v.compareTo(REPLAY_PROTECTED_V_MIN) > 0) {
-      chainId = Optional.of(v.subtract(REPLAY_PROTECTED_V_BASE).divide(TWO));
-      recId = v.subtract(TWO.multiply(chainId.get()).add(REPLAY_PROTECTED_V_BASE)).byteValueExact();
-    } else {
-      throw new RuntimeException(
-          String.format("An unsupported encoded `v` value of %s was found", v));
-    }
-    final BigInteger r = input.readUInt256Scalar().toBytes().toUnsignedBigInteger();
-    final BigInteger s = input.readUInt256Scalar().toBytes().toUnsignedBigInteger();
-    final SECP256K1.Signature signature = SECP256K1.Signature.create(r, s, recId);
-
-    input.leaveList();
-    chainId.ifPresent(builder::chainId);
-    return builder.signature(signature).build();
-  }
-
-  private static boolean isGoQuorumPrivateTransaction(final BigInteger v) {
-    return v.equals(GO_QUORUM_PRIVATE_TRANSACTION_V_VALUE_MAX)
-        || v.equals(GO_QUORUM_PRIVATE_TRANSACTION_V_VALUE_MIN);
-  }
+  Transaction decode(RLPInput rlpInput);
 }
